+++ src/modules/transfer/transfer.module.ts
@@ -1,0 +1,15 @@
+import { ContainerModule } from "inversify"
+import { TransferController } from "./transfer.controller"
+import { TransferRepository } from "./transfer.repository"
+import { TransferServiceImpl } from "./transfer.service"
+import {
+	type ITransferRepository,
+	type ITransferService,
+	TRANSFER_TYPES,
+} from "./transfer.types"
+
+export const TransferModule = new ContainerModule((bind) => {
+	bind<ITransferService>(TRANSFER_TYPES.TransferService).to(TransferServiceImpl)
+	bind<ITransferRepository>(TRANSFER_TYPES.TransferRepository).to(TransferRepository)
+	bind(TransferController).toSelf()
+})